Added tests for Db\Profiler.
